Features + Initiatives
Now displays features and initiatives.  Starting from the current
project, we're fetching all user stories at that level and below
(assuming the default profile settings aren't changed), and we're
fetching all features and initiatives at that level and below AND above.
We'll want to look at how to filter out some of the noise in the
future...